* optimizations / core cleanup (damn, how many places have uint64 for player guids, ffs)
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2038 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/EventableObject.cpp
@@ -376,7 +376,7 @@ uint32 EventableObject::event_GetEventPeriod(uint32 EventType)
 	m_lock.Acquire();
 	EventMap::iterator itr = m_events.find(EventType);
 	if(itr != m_events.end())
-		ret = itr->second->msTime;
+		ret = (uint32)itr->second->msTime;
 	
 	m_lock.Release();
 	return ret;
